Fix return value documentation (1/2) (#63477)
* Fix return value documentation (1/2).
* Avoid validation errors for missing 'contains:'.

+++ lib/ansible/modules/cloud/azure/azure_rm_automationaccount_info.py
@@ -202,7 +202,7 @@ automation_accounts:
                 limit:
                     description:
                         - Max limit, C(-1) for unlimited.
-                    type: long
+                    type: int
                     returned: always
                     sample: -1
                 name:
